Move users router in sub-module
