+++ fastapi_users/router/__init__.py
@@ -1,0 +1,2 @@
+from fastapi_users.router.common import ErrorCode, Event  # noqa: F401
+from fastapi_users.router.users import UserRouter, get_user_router  # noqa: F401
