+++ fastapi_users/router/common.py
@@ -1,0 +1,12 @@
+from enum import Enum, auto
+
+
+class ErrorCode:
+    REGISTER_USER_ALREADY_EXISTS = "REGISTER_USER_ALREADY_EXISTS"
+    LOGIN_BAD_CREDENTIALS = "LOGIN_BAD_CREDENTIALS"
+    RESET_PASSWORD_BAD_TOKEN = "RESET_PASSWORD_BAD_TOKEN"
+
+
+class Event(Enum):
+    ON_AFTER_REGISTER = auto()
+    ON_AFTER_FORGOT_PASSWORD = auto()
